matched func_8000B714_C314 2 attempts

+++ src/1E64A0.c
@@ -4,7 +4,7 @@
 
 extern void func_800B9440_1E64F0(void *);
 extern void func_800585C8_591C8(s32);
-extern void *func_8000B714_C314(void *, s16, s16);
+extern void *func_8000B714_C314(void *, s32, s32);
 extern void *func_80009EF8_AAF8(s32);
 extern void *func_80035F80_36B80(s32);
 extern void *func_8000B6B8_C2B8(s32);
